+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,6 +1,7 @@
 package nextstep.mvc;
 
 import nextstep.mvc.tobe.exception.RequestUrlNotFoundException;
+import nextstep.mvc.tobe.handlermapping.HandlerMapping;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
