+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handlermapping/AnnotationHandlerMapping.java
@@ -1,7 +1,10 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.handlermapping;
 
 import com.google.common.collect.Maps;
-import nextstep.mvc.ModelAndViewHandlerMapping;
+import nextstep.mvc.tobe.HandlerExecution;
+import nextstep.mvc.tobe.HandlerKey;
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.handlermapping.ModelAndViewHandlerMapping;
 import nextstep.mvc.tobe.exception.DuplicateRequestMappingException;
 import nextstep.mvc.tobe.exception.RenderFailedException;
 import nextstep.mvc.tobe.exception.RequestUrlNotFoundException;
