+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handlermapping/ControllerHandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.tobe.handlermapping;
 
 import nextstep.mvc.asis.Controller;
